pyproject.toml
[tool.poetry]
name = "pyserver"
version = "0.1.0"
description = ""
authors = ["Takehiro Suzuki <suzuki.tkw@om.asahi-kasei.co.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
asyncua = "^0.9.8"
pyserial = "^3.5"
flake8 = "^3.8.4"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
